19/07 Enemies Update
Enemies now shoot at you, but they only deal damage sometimes

+++ Assets/NewInventory/IModifiers.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IModifiers
 {
-    void AddValue(ref int baseValue);
+    void AddValue(ref float baseValue);
 }
